+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - PropertyGrid: Only show properties that are adorned with Browsable(true) if Browsable(false) not used #278
 - PropertyGrid: Add Button Control for properties inheriting from ICommand #279
 - PropertyGrid: Enhances options for PropertyGrid size #277
+- PropertyGrid: Add IValueConverter and ToolTip lookups for property grid array/collection elements #285
 
 ## [3.1.0]
 ### Added